# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REACTIFY API",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/cars": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Get all cars by pagination",
                "operationId": "e281fb4d4439d19a0326fe1cbd7eaffc",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/api/v1/brands": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Get all brands",
                "operationId": "41ee8c5ca737e71782e6d058a4c4a36f",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Hyundai"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/brands/{id}/models": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Get all models by brand id",
                "operationId": "8a95da88d6a983e1ae2aed71a8ce85ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the brand",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Hyundai"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/cars/{id}": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Get car by id",
                "operationId": "cee0c4406cfbf35c7b8733e6f65a1825",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the car",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 21
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "brand": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": "2"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Volkswagen"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "car_models": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": "5"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Passat"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "price": {
                                                        "type": "string",
                                                        "format": "float",
                                                        "example": "709.89"
                                                    },
                                                    "availability": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "year": {
                                                        "type": "integer",
                                                        "example": "2024"
                                                    },
                                                    "fuel_type": {
                                                        "type": "string",
                                                        "example": "Electric"
                                                    },
                                                    "transmission": {
                                                        "type": "string",
                                                        "example": "Manual"
                                                    },
                                                    "seats": {
                                                        "type": "integer",
                                                        "example": "5"
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "white"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "car1.jpg"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "In et enim laborum. Iste ex suscipit et tenetur et."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found"
                    }
                }
            }
        }
    }
}